pycharm_project_147/requirements.txt
absl-py==1.0.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blobfile==2.1.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.7
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.14.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.33.3
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
jax==0.4.13
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.3.4
jupyterlab-language-pack-zh-CN==3.3.post2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
keras==2.12.0
kiwisolver==1.4.2
libclang==16.0.6
lightning-bolts==0.7.0
lightning-utilities==0.10.1
llvmlite==0.41.1
lpips==0.1.4
lxml==4.9.4
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
ml-dtypes==0.2.0
mpi4py==3.1.5
mpmath==1.3.0
multidict==6.0.5
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
notebook==6.4.11
notebook-shim==0.1.0
numba==0.58.1
numpy==1.22.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.450.51
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
nvitop==0.5.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilgram==1.2.1
Pillow==9.1.0
piq==0.8.0
piqa==1.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.20.0
Pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytorch-lightning-bolts==0.3.2.post1
pytz==2022.1
PyYAML==6.0.1
pyzmq==22.3.0
regex==2023.12.25
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==2.2.1
torchaudio==0.11.0+cu113
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
triton==2.2.0
typing_extensions==4.10.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
wrapt==1.14.1
xgboost==1.6.0
xmltodict==0.12.0
yarl==1.9.4
zipp==3.8.0